update to support chef-11 upgrades to chef-12/13
