+++ .foodcritic
@@ -1,0 +1,1 @@
+~FC066
