+++ .kitchen.dokken.yml
@@ -1,0 +1,48 @@
+---
+# TAS50: THIS IS NOT A TYPICAL STANDARD DOKKEN KITCHEN FILE
+driver:
+  name: dokken
+  privileged: true # because Docker and SystemD/Upstart
+  chef_version: current
+
+transport:
+  name: dokken
+
+provisioner:
+  name: dokken
+  #chef_version: latest
+  #name: chef_zero
+  #root_path: /opt/kitchen
+  #require_chef_omnibus: '<%= ENV["OMNIBUS_CHEF_VERSION"] || "11.18.12" %>'
+  #  data_path: test/fixtures
+  #retry_on_exit_code:
+  #  - 213
+  #max_retries: 1
+  #wait_for_retry: 1
+  #client_rb:
+  #  exit_status: :enabled
+  #  client_fork: false
+
+verifier:
+  name: inspec
+
+platforms:
+  - name: ubuntu-16.04
+    driver:
+      image: ubuntu:16.04
+      pid_one_command: /bin/systemd
+      intermediate_instructions:
+        - RUN /usr/bin/apt-get update
+        - RUN /usr/bin/apt-get install apt-transport-https lsb-release procps net-tools -y
+
+suites:
+  - name: default
+    run_list:
+      - recipe[test::default]
+  - name: kill
+    run_list:
+      - recipe[test::kill]
+  - name: direct_url
+    run_list:
+      - recipe[test::direct_url]
+    includes: centos-6.9
