+++ .kitchen.yml
@@ -1,9 +1,9 @@
 driver:
   name: vagrant
-  require_chef_omnibus: '12.13.37'
+  require_chef_omnibus: '<%= ENV["OMNIBUS_CHEF_VERSION"] || "12.13.37" %>'
 
 provisioner:
-  name: chef_zero
+  name: chef_solo
   retry_on_exit_code:
     - 213
   max_retries: 1
