+++ Gemfile
@@ -1,9 +1,3 @@
-# This gemfile provides additional gems for testing and releasing this cookbook
-# It is meant to be installed on top of ChefDK which provides the majority
-# of the necessary gems for testing this cookbook
-#
-# Run 'chef exec bundle install' to install these dependencies
-
 source 'https://rubygems.org'
 
 gem 'tomlrb'
