+++ gemfiles/Gemfile.ancient
@@ -1,0 +1,11 @@
+source 'https://rubygems.org'
+gem 'chef', '~> 11.2.0'
+gem 'chefspec', '~> 3.4'
+gem 'rspec', '~> 2.14.1'
+gem 'berkshelf', '~> 3.0'
+gem 'nio4r', '~> 1.0'
+gem 'buff-ignore', '= 1.1.1'
+gem 'dep_selector', '= 1.0.3'
+gem 'ohai', '~> 6.0'
+gem 'mixlib-shellout', '~> 1.0'
+gem 'hashie', '~> 2.0'
