+++ gemfiles/Gemfile.ancient.lock
@@ -1,0 +1,161 @@
+GEM
+  remote: https://rubygems.org/
+  specs:
+    addressable (2.3.8)
+    berkshelf (3.3.0)
+      addressable (~> 2.3.4)
+      berkshelf-api-client (~> 1.2)
+      buff-config (~> 1.0)
+      buff-extensions (~> 1.0)
+      buff-shell_out (~> 0.1)
+      celluloid (~> 0.16.0)
+      celluloid-io (~> 0.16.1)
+      cleanroom (~> 1.0)
+      faraday (~> 0.9.0)
+      httpclient (~> 2.6.0)
+      minitar (~> 0.5.4)
+      octokit (~> 3.0)
+      retryable (~> 2.0)
+      ridley (~> 4.0)
+      solve (~> 1.1)
+      thor (~> 0.19)
+    berkshelf-api-client (1.3.1)
+      faraday (~> 0.9.1)
+      httpclient (~> 2.6.0)
+    buff-config (1.0.1)
+      buff-extensions (~> 1.0)
+      varia_model (~> 0.4)
+    buff-extensions (1.0.0)
+    buff-ignore (1.1.1)
+    buff-ruby_engine (0.1.0)
+    buff-shell_out (0.2.0)
+      buff-ruby_engine (~> 0.1.0)
+    celluloid (0.16.0)
+      timers (~> 4.0.0)
+    celluloid-io (0.16.2)
+      celluloid (>= 0.16.0)
+      nio4r (>= 1.1.0)
+    chef (11.2.0)
+      erubis
+      highline (>= 1.6.9)
+      json (~> 1.7.6, >= 1.4.4)
+      mixlib-authentication (>= 1.3.0)
+      mixlib-cli (~> 1.3.0)
+      mixlib-config (>= 1.1.2)
+      mixlib-log (>= 1.3.0)
+      mixlib-shellout
+      net-ssh (~> 2.6)
+      net-ssh-multi (~> 1.1.0)
+      ohai (>= 0.6.0)
+      rest-client (>= 1.0.4, < 1.7.0)
+      yajl-ruby (~> 1.1)
+    chefspec (3.4.0)
+      chef (~> 11.0)
+      fauxhai (~> 2.0)
+      rspec (~> 2.14)
+    cleanroom (1.0.0)
+    dep-selector-libgecode (1.3.1)
+    dep_selector (1.0.3)
+      dep-selector-libgecode (~> 1.0)
+      ffi (~> 1.9)
+    diff-lcs (1.3)
+    erubis (2.7.0)
+    faraday (0.9.2)
+      multipart-post (>= 1.2, < 3)
+    fauxhai (2.3.0)
+      net-ssh
+      ohai
+    ffi (1.9.18)
+    hashie (2.1.2)
+    highline (1.7.8)
+    hitimes (1.2.5)
+    httpclient (2.6.0.1)
+    ipaddress (0.8.3)
+    json (1.7.7)
+    mime-types (1.25.1)
+    minitar (0.5.4)
+    mixlib-authentication (1.4.1)
+      mixlib-log
+    mixlib-cli (1.3.0)
+    mixlib-config (2.2.4)
+    mixlib-log (1.7.1)
+    mixlib-shellout (1.6.1)
+    multipart-post (2.0.0)
+    net-ssh (2.9.4)
+    net-ssh-gateway (1.3.0)
+      net-ssh (>= 2.6.5)
+    net-ssh-multi (1.1)
+      net-ssh (>= 2.1.4)
+      net-ssh-gateway (>= 0.99.0)
+    nio4r (1.2.1)
+    octokit (3.8.0)
+      sawyer (~> 0.6.0, >= 0.5.3)
+    ohai (6.24.2)
+      ipaddress
+      mixlib-cli
+      mixlib-config
+      mixlib-log
+      mixlib-shellout
+      systemu (~> 2.5.2)
+      yajl-ruby
+    rest-client (1.6.9)
+      mime-types (~> 1.16)
+    retryable (2.0.4)
+    ridley (4.4.1)
+      addressable
+      buff-config (~> 1.0)
+      buff-extensions (~> 1.0)
+      buff-ignore (~> 1.1)
+      buff-shell_out (~> 0.1)
+      celluloid (~> 0.16.0)
+      celluloid-io (~> 0.16.1)
+      erubis
+      faraday (~> 0.9.0)
+      hashie (>= 2.0.2, < 4.0.0)
+      httpclient (~> 2.6)
+      json (>= 1.7.7)
+      mixlib-authentication (>= 1.3.0)
+      retryable (~> 2.0)
+      semverse (~> 1.1)
+      varia_model (~> 0.4.0)
+    rspec (2.14.1)
+      rspec-core (~> 2.14.0)
+      rspec-expectations (~> 2.14.0)
+      rspec-mocks (~> 2.14.0)
+    rspec-core (2.14.8)
+    rspec-expectations (2.14.5)
+      diff-lcs (>= 1.1.3, < 2.0)
+    rspec-mocks (2.14.6)
+    sawyer (0.6.0)
+      addressable (~> 2.3.5)
+      faraday (~> 0.8, < 0.10)
+    semverse (1.2.1)
+    solve (1.2.1)
+      dep_selector (~> 1.0)
+      semverse (~> 1.1)
+    systemu (2.5.2)
+    thor (0.19.4)
+    timers (4.0.4)
+      hitimes
+    varia_model (0.4.1)
+      buff-extensions (~> 1.0)
+      hashie (>= 2.0.2, < 4.0.0)
+    yajl-ruby (1.3.0)
+
+PLATFORMS
+  ruby
+
+DEPENDENCIES
+  berkshelf (~> 3.0)
+  buff-ignore (= 1.1.1)
+  chef (~> 11.2.0)
+  chefspec (~> 3.4)
+  dep_selector (= 1.0.3)
+  hashie (~> 2.0)
+  mixlib-shellout (~> 1.0)
+  nio4r (~> 1.0)
+  ohai (~> 6.0)
+  rspec (~> 2.14.1)
+
+BUNDLED WITH
+   1.11.2
